% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_function.R
\name{basehaz.est}
\alias{basehaz.est}
\title{Breslow estimator for estimating the baseline hazard}
\usage{
basehaz.est(fx, time, delta, t = NULL)
}
\arguments{
\item{fx}{vector of estimated hazard ratio}

\item{time}{vector of outcome}

\item{delta}{vector of censoring indicator}

\item{t}{time points of interest (default is NULL, using all observed survival time point)}
}
\value{
an object with attributes:
\itemize{
\item H0: vector of estimates of baseline hazard function
\item time.interest: vector of time point
}
}
\description{
Breslow estimator for estimating the baseline hazard
}
