% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_ensemble.R
\name{glasso}
\alias{glasso}
\title{Perform (adaptive) group lasso}
\usage{
glasso(X, y, group, nfolds = 5, adaptive = FALSE)
}
\arguments{
\item{X}{matrix of covariates}

\item{y}{vector of outcome}

\item{group}{vector of group membership}

\item{nfolds}{number of cross validation folds}

\item{adaptive}{logical: should adaptive group lasso be performed}
}
\value{
an object with attributes:
\itemize{
\item beta: coefficients
\item model: fitted model
\item time: computation time
\item best.lambda: Optimized tuning parameter lambda
\item adpen: adaptive weight
}
}
\description{
Perform (adaptive) group lasso
}
