% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_function.R
\name{is_almost_eq}
\alias{is_almost_eq}
\title{Check near equality (Reference from 'pre' package, Fokkema et al. 2020)}
\usage{
is_almost_eq(x, y, tolerance = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{value 1}

\item{y}{value 2}

\item{tolerance}{The tolerated difference between value 1 and value 2}
}
\value{
logical: If value 1 and value 2 are similar.
}
\description{
Check near equality (Reference from 'pre' package, Fokkema et al. 2020)
}
