% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_function.R
\name{pre_rules0}
\alias{pre_rules0}
\title{Reconstruct the data for rule ensemble procedure/prediction}
\usage{
pre_rules0(
  dat.test,
  rules.main,
  rules.treat,
  win = 0.025,
  x_scale = NULL,
  treat.index = NULL
)
}
\arguments{
\item{dat.test}{Input data frame (First column is observed survival time; Second column is censoring indicator; Third column is treatment indicator; The other columns are covariates)}

\item{rules.main}{set of main effect rules}

\item{rules.treat}{set of treatment effect rules}

\item{win}{winsorized rate (Default is 0.025)}

\item{x_scale}{normalized scale (Default is NULL, unspecified)}

\item{treat.index}{vector of specified treatment indicator (Default is NULL, similar to the input data)}
}
\value{
an object with attributes:
\itemize{
\item Y: vector of outcome
\item D: vector of censoring indicator
\item W: vector of treatment indicator
\item X: matrix of covariates
\item g_index: vector of group memebership
\item x_scale: vector of normalized scale for linear terms
\item treat: set of treatment effect rules
\item main: set of main effect rules
}
}
\description{
Reconstruct the data for rule ensemble procedure/prediction
}
