% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_survCRE.R
\name{predict_survCRE}
\alias{predict_survCRE}
\title{Predict function for proposed method}
\usage{
predict_survCRE(rule_ensemble, dat.test, time = -1, type = "survival rate")
}
\arguments{
\item{rule_ensemble}{An object from proposed method}

\item{dat.test}{The data used to estimate HTE}

\item{time}{Specify the time point to define the HTE, default is -1, the 0.8 quantile of observed survival time is used}

\item{type}{Type of HTE (survival rate or RMST)}
}
\value{
a vector of HTE
}
\description{
Predict function for proposed method
}
