% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{prop_imp}
\alias{prop_imp}
\title{Base function importance & Variable importance}
\usage{
prop_imp(mod.hat, dat)
}
\arguments{
\item{mod.hat}{a object created from function survCRE()}

\item{dat}{In put dataset}
}
\value{
an object with attributes:
\itemize{
\item base: base function importance
\item var: variable importance
}
}
\description{
Base function importance & Variable importance
}
