% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_function.R
\name{remove_complement}
\alias{remove_complement}
\title{Remove complement rules from the rules candidate (Reference from "pre" package, Fokkema et al., 2020)}
\usage{
remove_complement(dat, rules)
}
\arguments{
\item{dat}{Input data frame (First column is observed survival time; Second column is censoring indicator; Third column is treatment indicator; The other columns are covariates)}

\item{rules}{set of rules}
}
\value{
an object with attributes:
\itemize{
\item rules: set of rules
\item rulevars: binary matrix for rules
}
}
\description{
Remove complement rules from the rules candidate (Reference from "pre" package, Fokkema et al., 2020)
}
