% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_ensemble.R
\name{rule_cut}
\alias{rule_cut}
\title{Rule Ensemble procedure}
\usage{
rule_cut(pre_dat, rules.main, rules.treat, nfolds = 5)
}
\arguments{
\item{pre_dat}{a object created from sub function "pre_dat0"}

\item{rules.main}{set of main effect rules}

\item{rules.treat}{set of treatment effect rules}

\item{nfolds}{number of cross validation folds}
}
\value{
an object with attributes:
\itemize{
\item mod: Fitted model
\item lambda: Optimal lambda value
\item x_scale: Scale of features
\item rules.treat: Treatment rules
\item rules.main: Main effect rules
\item beta.hat: Coefficients
\item basehaz: Baseline hazard estimates
}
}
\description{
Rule Ensemble procedure
}
