% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_generation.R
\name{rule_gen}
\alias{rule_gen}
\title{Rule generation procedure}
\usage{
rule_gen(dat, depth = 2, learnrate = 0.01, ntrees = 333, sampfrac = NULL)
}
\arguments{
\item{dat}{Input data frame (First column is observed survival time; Second column is censoring indicator; Third column is treatment indicator; The other columns are covariates)}

\item{depth}{The mean depth of each tree-base function (Default is 2)}

\item{learnrate}{Shrinkage rate for each boosting step (Default is 0.01)}

\item{ntrees}{Number of trees (Default is 333)}

\item{sampfrac}{The fraction size of training sample for rule generation  (Default is NULL, means min(n / 2, 100 + 6 * sqrt(n), where n is the sample size)}
}
\value{
an object with attributes:
\itemize{
\item main: main effect rules
\item treat: treatment effect rules
}
}
\description{
Rule generation procedure
}
