% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCRE.R
\name{survCRE}
\alias{survCRE}
\title{Survival Causal Rule Ensemble}
\usage{
survCRE(
  dat,
  meandepth = 2,
  learnrate = 0.01,
  ntrees = 333,
  sampfrac = NULL,
  win = 0.025,
  nfolds = 5
)
}
\arguments{
\item{dat}{Input data frame (First column is observed survival time; Second column is censoring indicator; Third column is treatment indicator; The other columns are covariates)}

\item{meandepth}{The mean depth of each tree-base function (Default is 2)}

\item{learnrate}{Shrinkage rate for each boosting step (Default is 0.01)}

\item{ntrees}{Number of trees (Default is 333)}

\item{sampfrac}{The fraction size of training sample for rule generation  (Default is NULL, means min(n / 2, 100 + 6 * sqrt(n), where n is the sample size)}

\item{win}{winsorized rate (Default is 0.025)}

\item{nfolds}{number of cross validation folds}
}
\value{
an object with attributes:
\itemize{
\item main: main effect rules
\item treat: treatment effect rules
}
}
\description{
Survival Causal Rule Ensemble
}
\examples{
## Fit proposed method to ACTG175 dataset
library(speff2trial)

## Input dataset
data(ACTG175)
data <- ACTG175
data <- data[data$cd40>= 200 & data$cd40<= 500,]
data <- data[data$arms == 0|data$arms == 1,]

## Reconstruct dataset for proposed method
y <- data$days 
d <- data$cens 
z <- data$arms 
x.con <- data[,c("cd40","cd80","age","wtkg","karnof")] 
x.con <- data.frame(apply(x.con,2,as.numeric))
x.bin <- data[,c("hemo","homo","drugs","race","gender","str2","symptom")] 
x <- cbind(x.con,x.bin)
dat <- data.frame(Y = y, D = d, Z = z, x) # Construct the dataframe for analysis

## Fit the proposed method to dataset
fit <- survCRE(dat)

## Predict the HTE at the 0.8 quantile of observed surival time
pred <- predict_survCRE(fit, dat.test = dat)

}
