% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_function.R
\name{tran_rules}
\alias{tran_rules}
\title{Transform rules in the dataset into binary variables (Reference from "pre" package, Fokkema et al., 2020)}
\usage{
tran_rules(dat, rules)
}
\arguments{
\item{dat}{Input data frame (First column is observed survival time; Second column is censoring indicator; Third column is treatment indicator; The other columns are covariates)}

\item{rules}{set of rules}
}
\value{
binary matrix for rules
}
\description{
Transform rules in the dataset into binary variables (Reference from "pre" package, Fokkema et al., 2020)
}
